docs: reorganize files, add transition from bottom
