+++ docs/transition/transitions.stories.scss
@@ -1,0 +1,3 @@
+.space-bottom {
+  margin-bottom: var(--space-20);
+}
