+++ src/styles/styles.scss
@@ -1,2 +1,3 @@
 @forward "helpers";
 @forward "variables";
+@forward "transitions";
