+++ src/styles/transitions/slide.scss
@@ -1,24 +1,62 @@
-@mixin slide-transition($property, $duration: 0.5s) {
-  .slide {
-    &-enter-active,
-    &-leave-active {
-      transition: $property $duration ease;
-    }
+.slide-from-right {
+  &-enter-active,
+  &-leave-active {
+    transition: transform 0.5s ease;
+  }
   
-    &-enter-from {
-      position: fixed;
-    }
+  &-enter-from {
+    position: fixed;
+  }
   
-    &-enter-from,
-    &-leave-to {
-      transform: translate3d(100%, 0, 0);
-      transition: $property $duration ease-in;
-  
-      [dir="rtl"] & {
-        transform: translate3d(-100%, 0, 0);
-      }
+  &-enter-from,
+  &-leave-to {
+    transform: translate3d(100%, 0, 0);
+    transition: transform 0.5s ease-in;
+
+    [dir="rtl"] & {
+      transform: translate3d(-100%, 0, 0);
     }
   }
 }
 
+.slide-from-left {
+  &-enter-active,
+  &-leave-active {
+    transition: transform 0.5s ease;
+  }
+  
+  &-enter-from {
+    position: fixed;
+  }
+  
+  &-enter-from,
+  &-leave-to {
+    transform: translate3d(-100%, 0, 0);
+    transition: transform 0.5s ease-in;
 
+    [dir="rtl"] & {
+      transform: translate3d(-100%, 0, 0);
+    }
+  }
+}
+
+.slide-from-bottom {
+  &-enter-active,
+  &-leave-active {
+    transition: transform 0.5s ease;
+  }
+  
+  &-enter-from {
+    position: fixed;
+  }
+  
+  &-enter-from,
+  &-leave-to {
+    transform: translate3d(0, 100%, 0);
+    transition: transform 0.5s ease-in;
+
+    [dir="rtl"] & {
+      transform: translate3d(0, 100%, 0);
+    }
+  }
+}
