--- src/utilities/transition/docs.mdx
@@ -1,26 +1,0 @@
-import { Meta } from '@storybook/blocks';
-
-<Meta title="Utilities/Transitions/Docs" />
-
-# Transitions
-Transitions encompass animations that aid users in comprehending the flow within a page.
-Our Component Library includes a collection of readily available transition effects, which can be utilized as is or customized to align with your specific requirements.
-
-Few our components use our transitions:
-- [Modal](?path=/docs/organisms-modal--docs)
-- [SidePanel](?path=/docs/organisms-sidepanel--docs)
-
-## How to use
-To use our transition, just surround your component with build-in Vue component ```Transition``` with correct name of animations.<br />
-You can check how to use ```Transition``` in <a href="https://vuejs.org/guide/built-ins/transition.html" target="_blank">Vue documentation</a>
-
-For our transitions, we have separate story where you can see how them work.
-
-## Changing transitions
-If you desire to modify an existing transition, you have the ability to override the classes responsible for the effects.
-Replace ```{transition-name}``` with proper ```name``` according to your needs.
-
-- ```.{transition-name}-enter-active```
-- ```.{transition-name}-enter-from```
-- ```.{transition-name}-leave-active```
-- ```.{transition-name}-leave-to```
