--- src/utilities/transition/fade-transition/fade-transition.mdx
@@ -1,20 +1,0 @@
-import { Meta, Canvas } from '@storybook/blocks';
-import * as stories from './fade-transition.stories.js';
-import { StartNewCheckup } from '../../../components/organisms/UiModal/UiModal.stories.js'
-
-<Meta of={stories} />
-
-# Fade transition
-This visual effect that smoothly changes the opacity of an element, creating a gradual appearance or disappearance.<br/>
-This transition effect can be applied to various elements, such as text, images, or entire sections of a web page, adding a subtle and visually appealing animation to the user interface.
-
-## How to use
-To use ```fade``` transition, just surround your component with ```Transition``` <a href="https://vuejs.org/guide/built-ins/transition.html" target="_blank">Vue build-in component</a> with attribute ```name="fade"```
-
-## Used
-- [Modal](?path=/docs/organisms-modal--docs)
-- [SidePanel](?path=/docs/organisms-sidepanel--docs)
-
-## Stories
-### Modal
-<Canvas of={StartNewCheckup}/>
